gcp-to-real-time/requirements.txt
jwt == 1.1.0
paho-mqtt == 1.5.1
google-cloud-pubsub == 2.2.0
google-cloud == 0.34.0
google-cloud-bigquery==2.6.2
google-cloud-core== 1.5.0
pytz==2018.7